extract set_channel_var_sync to ari_ helper
